# docs/static/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "v1",
    "title": "Bordo-Blog API",
    "description": "Blog Management API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://bordo-blog.herokuapp.com/api/v1",
      "description": "Web server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/users/profile": {
      "summary": "",
      "description": "",
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User's profile",
        "description": "User's profile",
        "operationId": "users.profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/paths/~1users~1logout/post/responses/200/content/application~1json/schema/allOf/0"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "default": 200
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "type": "object",
                              "allOf": [
                                {
                                  "$ref": "#/paths/~1posts~1create/post/responses/201/content/application~1json/schema/allOf/2/properties/data/allOf/0"
                                },
                                {
                                  "$ref": "#/components/requestBodies/register/content/application~1json/schema/allOf/0/allOf/0"
                                },
                                {
                                  "$ref": "#/paths/~1users~1upload-avatar/post/responses/200/content/application~1json/schema/allOf/2/properties/data/allOf/3"
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "last_login": {
                                      "title": "User Last Login",
                                      "description": "User Last Login",
                                      "type": "string",
                                      "format": "date-time"
                                    }
                                  }
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "email": {
                                      "title": "User email address",
                                      "description": "User email address",
                                      "type": "string",
                                      "format": "email"
                                    },
                                    "email_verified": {
                                      "title": "User's email is verified?",
                                      "description": "User's email is verified?",
                                      "type": "boolean",
                                      "default": "true"
                                    },
                                    "email_notification": {
                                      "title": "User is subscribed the email notification?",
                                      "description": "User's email is verified?",
                                      "type": "boolean",
                                      "default": "true"
                                    }
                                  }
                                }
                              ]
                            },
                            {
                              "type": "object",
                              "properties": {
                                "last_login": {
                                  "title": "User Last Login",
                                  "description": "User Last Login",
                                  "type": "string",
                                  "format": "date-time"
                                }
                              }
                            },
                            {
                              "type": "object",
                              "properties": {
                                "roles": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "_id": {
                                        "title": "Role MongoDb Id",
                                        "description": "Role MongoDb Id",
                                        "type": "string",
                                        "example": "62ed69679c045c97a63390a4"
                                      },
                                      "name": {
                                        "title": "Role name",
                                        "description": "Role name",
                                        "type": "string",
                                        "example": "Writer"
                                      }
                                    }
                                  }
                                }
                              }
                            },
                            {
                              "type": "object",
                              "properties": {
                                "plan": {
                                  "type": "object",
                                  "properties": {
                                    "_id": {
                                      "title": "Plan MongoDb Id",
                                      "description": "Plan MongoDb Id",
                                      "type": "string",
                                      "example": "62ed69679c045c97a63390a4"
                                    },
                                    "name": {
                                      "title": "Plan name",
                                      "description": "Plan name",
                                      "type": "string",
                                      "example": "Basic Plan"
                                    },
                                    "right_to_view": {
                                      "title": "Plan name",
                                      "description": "Plan name",
                                      "type": "number",
                                      "example": 50
                                    }
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1users~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1tags~1get-all/get/responses/401"
          },
          "500": {
            "$ref": "#/paths/~1users~1login/post/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}